Add flag for timestamp creation (#97)
* Add flag for timestamp creation

* Change to non nullable Pair

* IntelliJ code inspection updates.

* Update date time format for epoch

* Merge conflicts

* Pre merge commit.

* Unit and integration test updates following addition of 'timestamp_created_from' field.

* Whitespace.

+++ src/main/kotlin/HbaseClient.kt
@@ -21,14 +21,14 @@ open class HbaseClient(val connection: Connection, private val columnFamily: Byt
         var success = false
         var attempts = 0
         var exception: Exception? = null
-        while (!success && attempts < Config.Hbase.retryMaxAttempts.toInt()) {
+        while (!success && attempts < Config.Hbase.retryMaxAttempts) {
             try {
                 putVersion(table, key, body, version)
                 success = true
             }
             catch (e: Exception) {
-                val delay = if (attempts == 0) Config.Hbase.retryInitialBackoff.toLong()
-                else (Config.Hbase.retryInitialBackoff.toLong() * attempts * Config.Hbase.retryBackoffMultiplier.toFloat()).toLong()
+                val delay = if (attempts == 0) Config.Hbase.retryInitialBackoff
+                else (Config.Hbase.retryInitialBackoff * attempts * Config.Hbase.retryBackoffMultiplier.toFloat()).toLong()
                 logger.warn("Failed to put batch ${e.message}", "attempt_number", "${attempts + 1}",
                     "max_attempts", "${Config.Hbase.retryMaxAttempts}", "retry_delay", "$delay", "error_message", "${e.message}")
                 Thread.sleep(delay)
